Add precheck class
